Don't serialize entire data model in consoleLog
